backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
openai==1.3.7
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.18
redis==5.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
numpy==1.24.3
scikit-learn==1.3.2
python-multipart==0.0.6
aioredis==2.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
sentence-transformers==2.2.2
torch==2.1.0